// src/interfaces/Keys.ts
export default interface Keys {
	up?: Phaser.Input.Keyboard.Key;
	down?: Phaser.Input.Keyboard.Key;
	left?: Phaser.Input.Keyboard.Key;
	right?: Phaser.Input.Keyboard.Key;
	w?: Phaser.Input.Keyboard.Key;
	s?: Phaser.Input.Keyboard.Key;
	a?: Phaser.Input.Keyboard.Key;
	d?: Phaser.Input.Keyboard.Key;

	space?: Phaser.Input.Keyboard.Key;
}